requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements.in
#
aiohttp==3.7.4.post0
    # via discord.py
alembic==1.6.5
    # via dataset
async-timeout==3.0.1
    # via aiohttp
attrs==21.2.0
    # via aiohttp
banal==1.0.6
    # via dataset
chardet==4.0.0
    # via aiohttp
configargparse==1.5
    # via -r requirements.in
dataset==1.5.0
    # via -r requirements.in
discord.py==1.7.3
    # via -r requirements.in
greenlet==1.1.0
    # via sqlalchemy
idna==3.2
    # via yarl
mako==1.1.4
    # via alembic
markupsafe==2.0.1
    # via mako
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
python-dateutil==2.8.1
    # via alembic
python-editor==1.0.4
    # via alembic
six==1.16.0
    # via python-dateutil
sqlalchemy==1.4.19
    # via
    #   alembic
    #   dataset
typing-extensions==3.10.0.0
    # via aiohttp
yarl==1.6.3
    # via aiohttp
